Split up show and edit agenda items components
