+++ modules/meeting/app/components/_index.sass
@@ -1,4 +1,4 @@
-@import "./meeting_agenda_items/item_component/show_component.sass"
+@import "./meeting_agenda_items/item_component.sass"
 @import "./meeting_agenda_items/form_component.sass"
 @import "./meeting_sections/header_component.sass"
 @import "./meetings/side_panel/state_component.sass"
