+++ modules/meeting/app/components/meeting_agenda_items/form_component.html.erb
@@ -6,7 +6,7 @@
       url: @submit_path,
       namespace: "form_#{wrapper_uniq_by}"
     ) do |f|
-      grid_layout("op-meeting-agenda-item-form", tag: :div) do |grid|
+      grid_layout("op-meeting-agenda-item-form", tag: :div, pl: 3) do |grid|
         grid.with_area(:title) do
           case @type
           when :simple
