+++ modules/meeting/app/components/meeting_agenda_items/item_component.html.erb
@@ -1,5 +1,5 @@
 <%=
-  component_wrapper do
+  component_wrapper(:border_box_row, **wrapper_arguments) do
     grid_layout("op-meeting-agenda-item", tag: :div, id: "item-#{@meeting_agenda_item.id}") do |grid|
       if drag_and_drop_enabled?
         grid.with_area(:"drag-handle", tag: :div, classes: "hide-when-print") do
