+++ modules/meeting/app/components/meeting_agenda_items/item_component.sass
